Remove sync headers from module map
to avoid needing separate module maps for building with and without sync.

This required moving the RLMRealmConfiguration+RealmSync category
to RLMRealmConfiguration_Private.h. This should really be in
RLMSyncUtil_Private.h but we just removed that from the module map.
