+++ Realm/Realm.modulemap
@@ -18,8 +18,6 @@ framework module Realm {
         header "RLMRealm_Private.h"
         header "RLMResults_Private.h"
         header "RLMSchema_Private.h"
-        header "RLMSyncConfiguration_Private.h"
-        header "RLMSyncUtil_Private.h"
     }
 
     explicit module Dynamic {
